telegram_bot/requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
APScheduler==3.11.2
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.5
defusedxml==0.7.1
frozenlist==1.6.0
greenlet==3.2.3
idna==3.10
jira==3.10.5
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
propcache==0.3.1
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0
python-dotenv==1.1.0
qrcode==8.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.20.0
YooMoney==0.1.2
